pyproject.toml
[project]
name = "renegade-sdk"
version = "0.1.4"
description = "Python SDK for Renegade darkpool"
readme = "README.md"
requires-python = ">=3.13"
license = "Apache-2.0"
authors = [{ name = "Renegade", email = "joey@renegade.fi" }]
keywords = [
    "defi",
    "darkpool",
    "zero-knowledge",
    "mpc",
    "blockchain",
    "arbitrum",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.1",
    "cryptography>=42.0.2",
    "web3>=6.15.1",
    "python-dotenv>=1.0.1",
    "deprecated>=1.2.18",
]

[project.urls]
Homepage = "https://renegade.fi/"
Repository = "https://github.com/renegade-fi/python-sdk"

[tool.hatch.build.targets.wheel]
packages = ["renegade"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
